+++ src/db/db_error.rs
@@ -8,14 +8,16 @@ pub enum DbError {
     ConnectionError(String),
     VoteError(String),
     AddStationError(String),
+    IllegalOrderError(String),
 }
 
 impl Display for DbError {
     fn fmt(&self, f: &mut Formatter) -> Result {
         match *self {
-            DbError::ConnectionError(ref v) => write!(f, "{}", v),
-            DbError::VoteError(ref v) => write!(f, "{}", v),
-            DbError::AddStationError(ref v) => write!(f, "{}", v),
+            DbError::ConnectionError(ref v) => write!(f, "ConnectionError '{}'", v),
+            DbError::VoteError(ref v) => write!(f, "VoteError '{}'", v),
+            DbError::AddStationError(ref v) => write!(f, "AddStationError '{}'", v),
+            DbError::IllegalOrderError(ref v) => write!(f, "IllegalOrderError '{}'", v),
         }
     }
 }